Handle null layout params.
The child view may have been created in a custom adapter without assigning layout params to it. In this case the app will crash with a runtime exception:

java.lang.NullPointerException: Attempt to read from field 'int android.view.ViewGroup$MarginLayoutParams.leftMargin' on a null object reference

At line 153 of my fork when `childWidthSpec` is defined.

I think this patch is preferable to a crash scenario.